@@ -1,106 +1,80 @@
 {
   "name": "beet",
   "productName": "beet",
-  "version": "0.6.8",
+  "version": "0.8.0",
   "apiversion": 3,
   "description": "Beet is a stand-alone key-manager and signing app for blockchains like BitShares, Steem and EOS, as well as BTC, BinanceChain and others.",
-  "main": "src/index.js",
-  "scripts": {
-    "start": "electron-forge start",
-    "buildcss": "node-sass --include-path src/scss src/scss/beet.scss src/css/beet.css",
-    "package": "electron-forge package",
-    "make": "electron-forge make",
-    "publish": "electron-forge publish",
-    "lint": "echo \"No linting configured\"",
-    "postpackage": "node write_digest.js"
-  },
-  "keywords": [],
   "author": "clockworkgr",
   "license": "MIT",
-  "config": {
-    "forge": {
-      "make_targets": {
-        "win32": [
-          "nsis"
-        ],
-        "darwin": [
-          "zip",
-          "dmg"
-        ],
-        "linux": [
-          "deb",
-          "rpm"
-        ]
-      },
-      "electronPackagerConfig": {
-        "packageManager": "npm"
-      },
-      "electronWinstallerConfig": {
-        "name": "Beet"
-      },
-      "electronInstallerDebian": {},
-      "github_repository": {
-        "owner": "bitshares",
-        "name": "beet",
-        "draft": true
-      }
-    }
-  },
+  "homepage": "https://github.com/bitshares/beet",
+  "main": "app/background.js",
   "build": {
-    "nsis": {
-      "allowElevation": true,
-      "allowToChangeInstallationDirectory": true,
-      "oneClick": false,
-      "perMachine": false
-    }
+    "appId": "io.get-beet.beet",
+    "files": [
+      "app/**/*",
+      "node_modules/**/*",
+      "package.json"
+    ],
+    "directories": {
+      "buildResources": "resources"
+    },
+    "publish": null
+  },
+  "scripts": {
+    "postinstall": "electron-builder install-app-deps",
+    "start": "node build/start.js",
+    "release": "webpack --config=build/webpack.app.config.js --env=production && electron-builder"
   },
   "dependencies": {
-    "@binance-chain/javascript-sdk": "^2.8.0",
-    "@panter/vue-i18next": "^0.15.0",
+    "@binance-chain/javascript-sdk": "^2.16.0",
+    "@panter/vue-i18next": "^0.15.1",
     "@smokenetwork/smoke-js": "^0.1.2",
-    "@whaleshares/wlsjs": "^0.1.1",
-    "bitcoinjs-lib": "^4.0.3",
-    "bitsharesjs": "^3.0.0-rc3",
-    "bootstrap-vue": "^2.0.0-rc.11",
+    "@whaleshares/wlsjs": "^0.1.7",
+    "bitcoinjs-lib": "^5.1.6",
+    "bitsharesjs": "^3.1.1",
+    "bootstrap-vue": "^2.0.3",
     "chalk": "^2.4.2",
     "crypto-js": "^3.1.9-1",
     "dexie": "^2.0.4",
-    "electron-compile": "^6.4.4",
     "electron-devtools-installer": "^2.2.4",
-    "electron-squirrel-startup": "^1.0.0",
-    "elliptic": "^6.4.1",
+    "elliptic": "^6.5.1",
     "eosjs": "^20.0.0",
+    "eosjs-ecc": "^4.0.7",
     "fs-jetpack": "^2.2.2",
-    "i18next": "^14.0.1",
-    "i18next-node-fs-backend": "^2.1.1",
+    "i18next": "^17.1.0",
+    "i18next-node-fs-backend": "^2.1.3",
     "lzma": "^2.3.2",
-    "node-fetch": "^2.3.0",
-    "otpauth": "^3.2.2",
+    "otpauth": "^5.0.3",
     "readline": "^1.3.0",
-    "steem": "^0.7.4",
-    "typeface-rajdhani": "0.0.54",
-    "typeface-roboto": "0.0.54",
-    "uuid": "^3.3.2",
-    "vue": "^2.5.22",
-    "vue-multiselect": "^2.1.3",
-    "vue-password-strength-meter": "^1.4.1",
-    "vue-router": "^3.0.2",
-    "vueify": "^9.4.1",
-    "vuex": "^3.1.0",
+    "steem": "^0.7.7",
+    "vue": "^2.6.10",
+    "vue-multiselect": "^2.1.6",
+    "vue-password-strength-meter": "^1.5.0",
+    "vue-router": "^3.1.3",
+    "vuex": "^3.1.1",
+    "ws": "^7.1.2",
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
-    "babel-eslint": "^10.0.1",
-    "babel-plugin-transform-async-to-generator": "^6.24.1",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "electron-forge": "^5.2.4",
-    "electron-forge-maker-nsis": "^20.37.0",
-    "electron-prebuilt-compile": "2.0.7",
-    "eslint": "^5.12.1",
-    "eslint-plugin-vue": "^5.2.2",
-    "hoek": ">=4.2.1",
-    "node-sass": "^4.11.0"
+    "@babel/core": "^7.6.3",
+    "@babel/plugin-proposal-object-rest-spread": "^7.6.2",
+    "@babel/plugin-transform-runtime": "^7.6.2",
+    "@babel/preset-env": "^7.6.3",
+    "babel-loader": "^8.0.6",
+    "css-loader": "^3.2.0",
+    "electron": "^3.0.0",
+    "electron-builder": "^21.2.0",
+    "file-loader": "^4.2.0",
+    "friendly-errors-webpack-plugin": "^1.7.0",
+    "node-sass": "^4.12.0",
+    "sass-loader": "^8.0.0",
+    "source-map-support": "^0.5.13",
+    "style-loader": "^1.0.0",
+    "vue-loader": "^15.7.1",
+    "vue-template-compiler": "^2.6.10",
+    "webpack": "^4.41.0",
+    "webpack-cli": "^3.3.9",
+    "webpack-merge": "^4.2.2",
+    "webpack-node-externals": "^1.7.2"
   }
 }
